requirements.txt
albumentations==1.3.1
altair==5.1.2
attrs==23.1.0
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1

efficientnet-pytorch==0.7.1
et-xmlfile==1.1.0
filelock==3.13.0
fonttools==4.43.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.18.0
idna==3.4
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.1
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
pdf2image==1.16.3
Pillow==10.0.1
#postgres==4.0
pdf2image==1.16.3
poppler-utils==0.1.0
pretrainedmodels==0.7.4
protobuf==4.24.4
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pyarrow==13.0.0
pyclipper==1.3.0.post5
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
segmentation-models-pytorch==0.3.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.2
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2

easyocr==1.7.1
spacy==3.7.2
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
#torch==2.0.1+cu117
#torchvision==0.15.2+cu117
